app/requirements.txt
flask==3.0.0
gunicorn==21.2.0
prometheus-client==0.19.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0

# Linting
flake8==6.1.0
black==23.12.0
